requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
build==0.10.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.27.8
packaging==23.1
PyJWT==2.7.0
pyproject-hooks==1.0.0
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
singlestoredb==0.8.1
sqlparams==5.1.0
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
